Scripted to Camera
Followed a tutorial on having the camera follow the player, and then referenced the controller script we learned in class to see if I could figure out how to get the mouse to control the camera orientation. It works but is maybe not the most efficient way to do it.